Fixed some bugs, fixed problem with primitive data types as arguments.

+++ SimpleCore/astring.h
@@ -25,15 +25,15 @@ public:
 
     AString arg(const AString& value);
 
-    AString left(uint n) const;
+    AString left(size_t n) const;
 
-    AString right(uint n) const;
+    AString right(size_t n) const;
 
     AString& append(char c);
 
     AString& append(const AString& string);
 
-    AString& fill(char c, int size = -1);
+    AString& fill(char c, size_t size = -1);
 
     AString& prepend(char c);
